requirements.txt
cffi==1.15.1
cryptg==0.4.0
pyaes==1.6.1
pyasn1==0.5.0
pycparser==2.21
rsa==4.9
Telethon==1.28.5
requests==2.30.0
python-dotenv==1.0.0
